+++ .gitignore
@@ -1,1 +1,3 @@
 __pycache__
+datasets
+.pytest_cache
